+++ bpdm-gate/src/test/kotlin/org/eclipse/tractusx/bpdm/gate/auth/AuthAdminIT.kt
@@ -63,7 +63,8 @@ class AuthAdminIT @Autowired constructor(
             getConfidenceCriteria = AuthExpectationType.Authorized
         ),
         uploadPartner = UploadPartnerAuthExpections(
-            postInput = AuthExpectationType.Authorized
+            postInput = AuthExpectationType.Authorized,
+            getInputTemplate = AuthExpectationType.Authorized
         )
     )
 ) {