+++ bpdm-gate/src/test/kotlin/org/eclipse/tractusx/bpdm/gate/auth/AuthOutputConsumerIT.kt
@@ -63,7 +63,8 @@ class AuthOutputConsumerIT @Autowired constructor(
             getConfidenceCriteria = AuthExpectationType.Authorized
         ),
         uploadPartner = UploadPartnerAuthExpections(
-            postInput = AuthExpectationType.Forbidden
+            postInput = AuthExpectationType.Forbidden,
+            getInputTemplate = AuthExpectationType.Forbidden
         )
     )
 ) {