+++ CHANGELOG.md
@@ -11,6 +11,7 @@ For changes to the BPDM Helm charts please consult the [changelog](charts/bpdm/C
 ### Added
 
 - BPDM Gate: Post endpoint to upload business partner input data using csv file.(#700)
+- BPDM Gate: GET endpoint to download the csv file template for business partner upload. (#700)
 
 ## [6.0.1] - [2024-05-27]